+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 26.1.15
 
+### Added
+
+- `mat-refined-prun-price`: Adds a "Refined PrUn Price" row
+
 ### Fixed
 
 - `XIT FINCH`: Fix the Y axis labels fraction digits